updated read me fixed repository types

+++ src/main/java/io/github/jperparas/resourcetrackerpwa/repositories/GpuRepository.java
@@ -4,5 +4,5 @@ import io.github.jperparas.resourcetrackerpwa.entities.Gpu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GpuRepository extends JpaRepository<Gpu, Long> {
+public interface GpuRepository extends JpaRepository<Gpu, Integer> {
 }
