+++ src/main/java/io/github/jperparas/resourcetrackerpwa/repositories/PowerTypeRepository.java
@@ -3,5 +3,5 @@ package io.github.jperparas.resourcetrackerpwa.repositories;
 import io.github.jperparas.resourcetrackerpwa.entities.PowerType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PowerTypeRepository extends JpaRepository<PowerType, Long> {
+public interface PowerTypeRepository extends JpaRepository<PowerType, Integer> {
 }
